Refactor codebase.split the code into different different files and folders. issue#33
